MDEV-12012/MDEV-11969 Can't remove GTIDs for a stale GTID Domain ID
As reported in MDEV-11969 "there's no way to ditch knowledge" about some
domain that is no longer updated on the master. Besides being of annoyance of
cluttering output in DBA console a stale domain can prevent the slave
to connect the master as MDEV-12012 witnesses.

This patch introduces a method to clean the server binlog state out
of obsolete gtid domains (which are estimated by the user as never receive
any update). The removal requires no event group from such domain are
present in existing binlog files though. If there is any it must be
first PURGEd in order for

  FLUSH BINARY LOGS DELETE_DOMAIN_ID=(list-of-domains)

to succeed. Otherwise it returns an error.

The list of obsolete domains can be computed through
intersecting two sets - the earliest (first) binlog's Gtid_list
and the current value of @@global.gtid_binlog_state - and extracting
the domain id component from each item the result list.
The list of obsolete ("unused") domains means there are no event group
from any of the domains in the current binlog files.
The new DELETE_DOMAIN_ID FLUSH rotates binlog to omit the deleted domains
in the new active binlog file's Gtid_list.
Notice though when the command is ineffective that is no requested to delete
domain exists rotation does not occur.

Obsolete domain deletion is not harmful for connected slaves as long
as master side binlog files *purge* is synchronized with them.
The slaves must have the last event from purged files processed as usual,
in order not to bump later into requesting a gtid from a file which
was already gone.
While the command is not replicated (as ordinary FLUSH BINLOG LOGS is)
slaves, even though having extra domains, won't suffer from reconnection errors
thanks to master-slave gtid connection protocol allowing the master
to be ignorant (and then remain to be that in future) about a gtid domain.
Should such slave to be promoted into master role at failure it may run
the ex-master's time

 FLUSH BINARY LOGS DELETE_DOMAIN_ID=(list-of-domains)

over again to clean its own binlog state.